refactor: rename feature-related components
